# Performance Optimization: Go Concurrency Improvements for VAINO