… docs/PERFORMANCE_OPTIMIZATIONS.md
… internal/storage/concurrent_benchmark_test.go
… internal/workers/diff_worker_benchmark_test.go
… test/performance/optimization_benchmarks_test.go